update tests file within tickets routes

+++ tickets/src/test/setup.ts
@@ -12,19 +12,23 @@ jest.mock('../nats-wrapper');
 
 let mongo: any;
 beforeAll(async () => {
-    jest.clearAllMocks();
-    process.env.JWT_KEY = 'asdf'
-    mongo = await MongoMemoryServer.create();
+    process.env.JWT_KEY = 'asdfasdf';
+    process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
+
+    const mongo = await MongoMemoryServer.create();
     const mongoUri = mongo.getUri();
 
     await mongoose.connect(mongoUri, {});
 });
 
 beforeEach(async () => {
-    const collections = await mongoose.connection.db.collections();
+    jest.clearAllMocks();
+    if (mongoose.connection.db) {
+        const collections = await mongoose.connection.db.collections();
 
-    for (let collection of collections) {
-        await collection.deleteMany({});
+        for (let collection of collections) {
+            await collection.deleteMany({});
+        }
     }
 });
 
@@ -36,15 +40,16 @@ afterAll(async () => {
 });
 
 global.signin = () => {
-    // Build a JWT payload. { id, email }
+    // Build a JWT payload.  { id, email }
     const payload = {
         id: new mongoose.Types.ObjectId().toHexString(),
-        email: 'test@test.com'
-    }
-    // Create the JWT
+        email: 'test@test.com',
+    };
+
+    // Create the JWT!
     const token = jwt.sign(payload, process.env.JWT_KEY!);
 
-    // Build session Object { jwt: MY_JWT}
+    // Build session Object. { jwt: MY_JWT }
     const session = { jwt: token };
 
     // Turn that session into JSON
@@ -53,6 +58,6 @@ global.signin = () => {
     // Take JSON and encode it as base64
     const base64 = Buffer.from(sessionJSON).toString('base64');
 
-    // return a string that is the cookie with the encoded data
+    // return a string thats the cookie with the encoded data
     return [`session=${base64}`];
 };
